fix: Rename shared limit client file
